+++ modules/project/core/src/sap/commerce/toolset/project/facet/YFacetEditorTab.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -58,7 +58,7 @@ class YFacetEditorTab(
 
             row {
                 label("Type:")
-                label(extensionDescriptor.type.name)
+                label(extensionDescriptor.type.title)
                 icon(extensionDescriptor.type.icon)
             }.layout(RowLayout.PARENT_GRID)
 
@@ -70,6 +70,12 @@ class YFacetEditorTab(
                         icon(it.icon)
                     }.layout(RowLayout.PARENT_GRID)
                 }
+
+            row {
+                label("Source availability:")
+                label(extensionDescriptor.sourceAvailability.title)
+                icon(extensionDescriptor.sourceAvailability.icon)
+            }.layout(RowLayout.PARENT_GRID)
         }
 
         if (extensionDescriptor.subModuleType == SubModuleDescriptorType.ADDON && extensionDescriptor.installedIntoExtensions.isNotEmpty()) {